remove duplicate data fusion label
